+++ CHANGELOG.md
@@ -2,6 +2,8 @@ v0.5.0
  - Implemented the `logistic::logistic` special function
  - Implemented the `logistic::logit` special function
  - Implemented the `factorial::multinomial` special function
+ - Implemented the `harmonic::harmonic` special function
+ - Implemented the `harmonic::gen_harmonic` special function
 
 v0.4.0
 - Implemented the `exponential::integral` special function